requirements.txt
# HTTP Requests
requests>=2.31.0

# HTML Parsing
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Text Processing
nltk>=3.8.0
regex>=2023.8.8

# Embeddings
sentence-transformers>=2.2.0
openai>=1.3.0
transformers>=4.35.0
torch>=2.1.0

# Vector Database
faiss-cpu>=1.7.4
numpy>=1.24.0

# API Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Utility
python-dotenv>=1.0.0
python-multipart>=0.0.6

# Logging and Development
loguru>=0.7.0
rich>=13.6.0

# Data Processing
pandas>=2.1.0
pathlib>=1.0.1

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0